refactor: rename query folder from gnosisdev to gnosis-ai
