+++ queries/gnosis-ai/index.ts
