+++ queries/gnosis-ai/queries.ts
